+++ test/request/header.js
@@ -9,4 +9,10 @@ describe('req.header', () => {
     const req = request();
     assert.deepEqual(req.header, req.req.headers);
   });
+
+  it('should set the request header object', () => {
+    const req = request();
+    req.header = {'X-Custom-Headerfield': 'Its one header, with headerfields'};
+    assert.deepEqual(req.header, req.req.headers);
+  });
 });